3.0 final changes
Final changes for 3.0
Fixed icon feature bug for other OS.
Changed the way macOS runs the exploit it now opens a new terminal and prompts for a password to make it more obvious what should be done after pressing "Run PPPwn"
Changed the way macOS handles interface names
A change i missed to mention in 3.0 beta is that the pppwn executible and other files are in the pppwn folder now so that people are not confused on which exe to run and to make it a bit tidier.

GUI now supports Windows 7 to 11:
Full 3.0 release will have 3 Windows zips.
Windows-7.
Windows-8-8.1.
Windows-10-11.
Linux and macOS support is also at a good place.
Full 3.0 release will be out in a few days and will have an OS specific readme in each zip alongside the updated readme for the repo which reflects the new changes to the way of running the GUI and compiling it.

GoldHEN for 9.03 is not yet released (only the stage2 bin)
HEN VTX received a small update which will be added alongside the next GoldHEN version release.
The next GUI version will be 3.1 which will have the new GoldHEN and VTX files.